+++ persons.json
@@ -1,2 +1,3 @@
 {"id":"1649871772.7865512","type":"Student","name":"s1","age":15,"parent_permission":true}
 {"id":"1649871783.9507306","type":"Teacher","name":"t2","age":45,"specialization":"sleep"}
+{"id":"1649872411.319786","type":"Teacher","name":"das","age":43,"specialization":"zxc"}
